Remove the surface slope limiter
The surface slope limiter may produce "better looking" velocity fields in diagnostic
computations but it is a disaster when coupled to mass transport.

The problem is that limiting the surface slope in the stress balance solver allows the
model to develop very steep cliffs since the stress balance model cannot respond to these
changes in ice geometry. We avoid oscillations in ice velocity but the ice geometry is
wrong...